functional delete button added
