+++ api/notes.json
@@ -17,13 +17,6 @@
       "dateOfNote": "10/19/2021",
       "noteText": "This is test 3",
       "suspect": "Taurean Bins"
-    },
-    {
-      "id": 4
-    },
-    {
-      "noteText": "Akiko Endo",
-      "id": 6
     }
   ]
 }
