+++ src/index.html
@@ -15,10 +15,6 @@
             <a id="criminals-nav-link" href="#">Criminals</a>
           </li>
 
-          <article ></article>
-         
-          
-
           <li>
             <a id="officers-nav-link" href="#">Officers</a>
           </li>
@@ -33,15 +29,22 @@
   
     <button id="dark-mode">Dark Mode</button>
 
+    <div class="filters__crime"></div><!--this is a crime menu-->
+    <div class="filters__officer"></div><!--this is a officer menu-->
+    
     <section class="noteFormContainer"></section>
     <section class="content-container"></section>
     <section class="criminalsContainer"></section>
 
     <div class="note-list"></div>
     <div class="print-list row"></div>
-    <div class="filters__crime"></div>
-    <div class="criminal-list"></div>
-    <div class="filters__officer"></div>
+    
+
+    <div class="criminal-list"></div><!--this is where the all crimals will printed from the nav bar-->
+
+    
+    <div class="officer-list"></div><!--this is where the all officers will printed from the nav bar-->
+
    <!-- <section class="criminals">
         
         <div class="convictions-select-container"></div>
