+++ src/scripts/notes/NoteCard.js
@@ -1,4 +1,5 @@
-
+import { deleteNote } from "./NoteDataProvider.js";
+import { NoteList } from "./NoteList.js";
 export const note = (taco) => {
     return `
     <div class="col-sm-3 border border-dark m-3 list-group-item">
@@ -11,3 +12,19 @@ export const note = (taco) => {
         `
 }
 
+const eventHub = document.querySelector(".content-container")
+
+eventHub.addEventListener("click", removeEvent => {
+  if (removeEvent.target.id.startsWith("deleteNote")) {
+    
+    const idToDelete = removeEvent.target.id.split("--")[1]
+    // ---------- Write your code here -------------//
+    // Call the deleteNote function and pass in the appropriate id
+    
+    console.log(idToDelete)
+    deleteNote (idToDelete)
+    .then(NoteList)
+    // Then call NoteList to refresh the list of notes
+
+  }
+});
